Move dumper creator SPI to api module
