+++ shardingsphere-kernel/shardingsphere-data-pipeline/shardingsphere-data-pipeline-api/src/main/java/org/apache/shardingsphere/data/pipeline/spi/ingest/dumper/InventoryDumperCreatorFactory.java
@@ -15,14 +15,13 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.data.pipeline.core.ingest.dumper;
+package org.apache.shardingsphere.data.pipeline.spi.ingest.dumper;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.apache.shardingsphere.infra.util.spi.ShardingSphereServiceLoader;
+import org.apache.shardingsphere.infra.util.spi.type.required.RequiredSPIRegistry;
 import org.apache.shardingsphere.infra.util.spi.type.typed.TypedSPIRegistry;
-
-import java.util.Optional;
 
 /**
  * Inventory dumper creator factory.
@@ -41,7 +40,6 @@ public class InventoryDumperCreatorFactory {
      * @return InventoryDumperCreator
      */
     public static InventoryDumperCreator getInstance(final String databaseType) {
-        Optional<InventoryDumperCreator> result = TypedSPIRegistry.findRegisteredService(InventoryDumperCreator.class, databaseType);
-        return result.orElseGet(DefaultInventoryDumperCreator::new);
+        return TypedSPIRegistry.findRegisteredService(InventoryDumperCreator.class, databaseType).orElseGet(() -> RequiredSPIRegistry.getRegisteredService(InventoryDumperCreator.class));
     }
 }
